+++ Contents/mods/B42 PZLinux/42/media/lua/client/Context/World/Features/PZLinuxDarkWeb.lua
@@ -779,6 +779,7 @@ function darkWebUI:onSellItem(button)
 end
 
 function darkWebUI:onHelp()
+    PZLinuxDarkWebLoadCustomItems()
     self.shoppingBuyButton:setVisible(false)
     self.shoppingSellButton:setVisible(false)
     self.shoppingHelpButton:setVisible(false)